myproject/requirements.txt
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
sqlparse==0.5.3
tzdata==2025.2